third_party/WebKit/LayoutTests/fast/forms/checkValidity-001-expected.txt

CONSOLE WARNING: The <keygen> element is deprecated and will be removed in M54, around October 2016. See https://www.chromestatus.com/features/5716060992962560 for more details.
This test checks if checkValidity() returns correctly a true (meaning no error) result.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 lorem ipsum     
PASS document.getElementsByTagName("fieldset")[0].checkValidity() is true
PASS document.getElementsByTagName("input")[0].checkValidity() is true
PASS document.getElementsByTagName("button")[0].checkValidity() is true
PASS document.getElementsByTagName("select")[0].checkValidity() is true
PASS document.getElementsByTagName("textarea")[0].checkValidity() is true
PASS document.getElementsByTagName("output")[0].checkValidity() is true
PASS document.getElementsByTagName("object")[0].checkValidity() is true
PASS document.getElementsByTagName("keygen")[0].checkValidity() is true
PASS successfullyParsed is true

TEST COMPLETE
